Only allow commanders to access their own guild
Cross-guild access is ok for viewing

+++ AnspiritConsoleUI/Commands/AnzacDiscordOfficerModule.cs
@@ -2,6 +2,7 @@
 using AnspiritConsoleUI.Constants;
 using AnspiritConsoleUI.Services;
 using AnspiritConsoleUI.Services.Database;
+using AnspiritConsoleUI.Utilities;
 using Discord;
 using Discord.Commands;
 using Discord.WebSocket;
@@ -22,35 +23,49 @@ namespace AnspiritConsoleUI.Commands
         [Summary("Adds an officers notes categories")]
         public async Task AddOfficerNote(IUser discordUser, string categoryName, [Remainder] string notes)
         {
-            await DbService.AddOfficerNote(categoryName, discordUser.Id, notes);
-            await ReplyNewEmbed($"Added the note successfully", Color.Green);
+            if (DiscordUtilities.UserIsInCallingOfficersGuild(Context.User as SocketGuildUser, discordUser as SocketGuildUser))
+            {
+                await DbService.AddOfficerNote(categoryName, discordUser.Id, notes);
+                await ReplyNewEmbed("Added the note successfully", Color.Green);
+            }
+            else
+            {
+                await ReplyNewEmbed("You cannot access members in other guilds.", Color.Red);
+            }
         }
         [Command("officernotes remove")]
         [Alias("officernotes delete", "officernotes del", "officernotes rem")]
         [Summary("Removes an officer note")]
         public async Task RemoveOfficerNote(IUser discordUser, string categoryName, [Summary("dd/mm/yyyy")] string dateTime)
         {
-            // TODO: This could be parsed simply by DateTime.Parse
-            var datetimeParts = dateTime.Split('/');
-            var dayString = datetimeParts[0];
-            var monthString = datetimeParts[1];
-            var yearString = datetimeParts[2];
+            if (DiscordUtilities.UserIsInCallingOfficersGuild(Context.User as SocketGuildUser, discordUser as SocketGuildUser))
+            {
+                // TODO: This could be parsed simply by DateTime.Parse
+                var datetimeParts = dateTime.Split('/');
+                var dayString = datetimeParts[0];
+                var monthString = datetimeParts[1];
+                var yearString = datetimeParts[2];
 
-            if (!int.TryParse(dayString, out var day))
-            {
-                throw new Exception("Could not parse the day from the datetime string, are you using dd/mm/yyyy format?");
-            }
-            if (!int.TryParse(monthString, out var month))
-            {
-                throw new Exception("Could not parse the month from the datetime string, are you using dd/mm/yyyy format?");
-            }
-            if (!int.TryParse(yearString, out var year))
-            {
-                throw new Exception("Could not parse the year from the datetime string, are you using dd/mm/yyyy format?");
-            }
+                if (!int.TryParse(dayString, out var day))
+                {
+                    throw new Exception("Could not parse the day from the datetime string, are you using dd/mm/yyyy format?");
+                }
+                if (!int.TryParse(monthString, out var month))
+                {
+                    throw new Exception("Could not parse the month from the datetime string, are you using dd/mm/yyyy format?");
+                }
+                if (!int.TryParse(yearString, out var year))
+                {
+                    throw new Exception("Could not parse the year from the datetime string, are you using dd/mm/yyyy format?");
+                }
 
-            await DbService.RemoveOfficerNotesAsync(discordUser.Id, categoryName, new DateTime(year, month, day));
-            await ReplyNewEmbed($"Removed the note successfully", Color.Green);
+                await DbService.RemoveOfficerNotesAsync(discordUser.Id, categoryName, new DateTime(year, month, day));
+                await ReplyNewEmbed($"Removed the note successfully", Color.Green);
+            }
+            else
+            {
+                await ReplyNewEmbed("You cannot access members in other guilds.", Color.Red);
+            }
         }
         [Command("officernotes list")]
         [Alias("officernotes ls", "officernotes get", "officernotes")]
